Refactor evaluation tests by behavior
